pyproject.toml
[tool.poetry]
name = "wolt"
version = "0.1.0"
description = ""
authors = ["raianul kabir <raianul.kabir@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
requests = "^2.28.1"
python-environ = "^0.4.54"
flask-marshmallow = "^0.14.0"
black = "^22.12.0"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
ipdb = "^0.13.11"
mypy = "^0.991"
pylint = "^2.15.9"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
